Merge dylan

+++ src/components/AccountComponents/UserInfo.js
@@ -1,23 +1,30 @@
-import React, { Component, useState } from 'react'
-import UserService from '../../services/UserService'
-import Button from 'react-bootstrap/Button'
-import Collapse from 'react-bootstrap/Collapse'
+import React, { Component, useState } from 'react';
+import UserService from '../../services/UserService';
+import Button from 'react-bootstrap/Button';
+import Collapse from 'react-bootstrap/Collapse';
+import jwt_decode from 'jwt-decode';
+import { withRouter } from "react-router";
+import { Redirect } from 'react-router-dom';
 
-export default class UserInfo extends Component {
+class UserInfo extends Component {
   constructor(props) {
     super(props);
     this.state = {
         user: {},
         user_loaded: false,
+        redirect: false,
     }
   }
+  //loads in user
+  //keeps user fields hidden when the user cannot be loaded
   componentDidMount() {
     try{
-      UserService.read().then(res => {
+      UserService.get().then(res => {
         this.setState({user: res.data});
         this.setState({user_loaded: true});
       }).catch((error)=>{
         console.log(error);
+        this.setState({user_loaded: false});
       });
     }
     catch(e){
@@ -25,19 +32,27 @@ export default class UserInfo extends Component {
     }
     
   }
-  closeAccount() {
-    try{
-      UserService.delete().then(resp => {
-        this.props.history.push('/login');
-      }).catch((e)=>{
-        console.log(e);
-      });
-    }
-    catch(e){
-      console.log(e);
-    }
-  }
   render (){
+    //exits to root page on successful closing of the account
+    function closeAccount() {
+      if(localStorage.getItem("jwt")!==null&&document.getElementById('confirmation').value===jwt_decode(localStorage.getItem("jwt")).sub){
+        try{
+          UserService.delete().then(resp => {
+            console.log('User account deleted');
+            localStorage.removeItem('jwt');
+            this.props.history.push('login');
+          }).catch((e)=>{
+            console.log(e);
+          });
+        }
+        catch(e){
+          console.log(e);
+        }
+      }
+      else{
+        document.getElementById('confirmation_mismatch').textContent='Username mismatch';
+      }
+    }
     function CloseConfirmation() {
       const [open, setOpen] = useState(false);
       return (
@@ -54,16 +69,21 @@ export default class UserInfo extends Component {
               <input id='confirmation' type='text'></input>
               <br/>
               <p1>Type your username to confirm the closing of your account</p1>
-              <button onClick={this.closeAccount}>submit</button>
+              <br/>
+              <p1 id='confirmation_mismatch' style={{color:'red'}}></p1>
+              <br/>
+              <button onClick={closeAccount}>submit</button>
             </div>
           </Collapse>
         </>
       );
     }
     return (
-    <div className="UserInfo">
+      <div>
+        {this.state.redirect?<Redirect push to='/login'/>:null}
         <h1>Welcome to the User Info </h1>
-        <Collapse in={this.state.user_loaded}>
+        {this.state.user_loaded?(
+          <div>
           <h2>Username: </h2>
           <p1 id='username'>{this.state.user.username}</p1>
           <h2>First Name:</h2>
@@ -75,12 +95,16 @@ export default class UserInfo extends Component {
           <h2>Phone:</h2>
           <p1 id='phone'>{this.state.user.phone}</p1>
           <div>
-            <a href='../update'>Update your Account</a>
+            <button onClick={()=>{this.props.history.push('update');}}>Update your Account</button>
           </div>
-        <CloseConfirmation/>
-        </Collapse>
-        
+          <div>
+            <CloseConfirmation/>
+          </div>
+          </div>
+        ):null}        
     </div>
     );
     }
   }
+
+  export default withRouter(UserInfo);
